change from float and int 64 to 32

+++ recbole/data/dataset/dataset.py
@@ -478,7 +478,7 @@ class Dataset(torch.utils.data.Dataset):
                     self.field2bucketnum[field] = 2
             columns.append(field)
             usecols.append(field_type)
-            dtype[field_type] = np.float64 if ftype == FeatureType.FLOAT else str
+            dtype[field_type] = np.float32 if ftype == FeatureType.FLOAT else str
 
         if len(columns) == 0:
             self.logger.warning(f"No columns has been loaded from [{source}]")
@@ -650,7 +650,7 @@ class Dataset(torch.utils.data.Dataset):
                 elif ftype == FeatureType.FLOAT:
                     feat[field].fillna(value=feat[field].mean(), inplace=True)
                 else:
-                    dtype = np.int64 if ftype == FeatureType.TOKEN_SEQ else np.float
+                    dtype = np.int32 if ftype == FeatureType.TOKEN_SEQ else np.float
                     feat[field] = feat[field].apply(
                         lambda x: np.array([], dtype=dtype)
                         if isinstance(x, float)
@@ -1470,7 +1470,7 @@ class Dataset(torch.utils.data.Dataset):
         """Get the average number of users' interaction records.
 
         Returns:
-            numpy.float64: Average number of users' interaction records.
+            numpy.float32: Average number of users' interaction records.
         """
         if isinstance(self.inter_feat, pd.DataFrame):
             return np.mean(self.inter_feat.groupby(self.uid_field).size())
@@ -1484,7 +1484,7 @@ class Dataset(torch.utils.data.Dataset):
         """Get the average number of items' interaction records.
 
         Returns:
-            numpy.float64: Average number of items' interaction records.
+            numpy.float32: Average number of items' interaction records.
         """
         if isinstance(self.inter_feat, pd.DataFrame):
             return np.mean(self.inter_feat.groupby(self.iid_field).size())
@@ -2018,7 +2018,7 @@ class Dataset(torch.utils.data.Dataset):
             row_num, max_col_num = self.item_num, self.user_num
             row_ids, col_ids = item_ids, user_ids
 
-        history_len = np.zeros(row_num, dtype=np.int64)
+        history_len = np.zeros(row_num, dtype=np.int32)
         for row_id in row_ids:
             history_len[row_id] += 1
 
@@ -2034,7 +2034,7 @@ class Dataset(torch.utils.data.Dataset):
                 f"{col_num / max_col_num * 100}% of the total."
             )
 
-        history_matrix = np.zeros((row_num, col_num), dtype=np.int64)
+        history_matrix = np.zeros((row_num, col_num), dtype=np.int32)
         history_value = np.zeros((row_num, col_num))
         history_len[:] = 0
         for row_id, value, col_id in zip(row_ids, values, col_ids):
